# back/src/swagger.yaml
openapi: 3.0.0
info:
  version: 0.2.0
  title: Pokemon-Collector API
  description: Pokemon-Collector API 명세서입니다
  license:
    name: MIT
servers:
  - url: 'http://localhost:5000/'
components:
  securitySchemes:
    Authorization:
      in: header
      name: Authorization
      type: apiKey
      scheme: bearer
...
paths:
  /user/register:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nickname:
                  type: string
                  example: 포켓몬 트레이너
                email:
                  type: string
                  example: a@naver.com
                password:
                  type: string
                  example: 1234
                sex:
                  type: string
                  example: Male
                birth:
                  type: string
                  example: 1990-12-03T00:00:00.000Z
                interest:
                  type: number
                  example: 3
                likeType:
                  type: string
                  example: 격투
      tags:
        - user
      summary: 회원가입
      description: User의 회원가입 API
      responses:
        '200':
          description: 회원가입 성공
        '400':
          description: Content-Type application/json으로 설정
  /user/login:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: a@naver.com
                password:
                  type: string
                  example: 1234
      tags:
        - user
      summary: 로그인
      description: User의 로그인 API
      responses:
        '200':
          description: 로그인 성공
        '400':
          description: 가입 내역이 없는 이메일이므로 로그인 실패
  /user/current:
    get:
      tags:
        - user
      summary: 최근에 로그인한 유저 정보
      security:
        - Authorization: []
      description: 최근에 로그인한 User의 정보를 불러오는 API
      responses:
        '200':
          description: 최근에 로그인한 유저 정보 불러옴
        '400':
          description: 정상적인 토큰이 아니라서 유저 정보를 불러오는 것을 실패
          content:
            application/json:
              schema:
                type: object
  /user/attendanceCheck:
    put:
      tags:
        - user
      summary: 유저의 마지막 로그인 시간으로부터 24시간 경과 확인
      description: 유저의 마지막 로그인 시간으로부터 24시간 경과 확인 -> 포인트 지급 여부 업데이트 API
      security:
        - Authorization: []
      responses:
        '200':
          description: '출석체크 시간, 포인트 지급 여부 업데이트'
          content:
            application/json:
              schema:
                type: object
                properties:
                  attendance:
                    type: string
                    example: 2022-05-02T16:03:30.429Z
                  isPointGiven:
                    type: string
                    example: false
  /user/checkIn:
    put:
      tags:
        - user
      summary: 기존 포인트 + 1000
      description: '기존 포인트 + 1000, 포인트 지급 여부 업데이트 API'
      security:
        - Authorization: []
      responses:
        '200':
          description: '포인트, 포인트 지급 여부 업데이트'
          content:
            application/json:
              schema:
                type: object
                properties:
                  attendance:
                    type: string
                    example: 2022-05-02T16:03:30.429Z
                  isPointGiven:
                    type: string
                    example: false
                  point:
                    type: number
                    example: 2000
  /user/profileModify:
    put:
      tags:
        - user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nickname:
                  type: string
                  example: 태초마을 관장
                likeType:
                  type: string
                  example: 불꽃
                profileImg:
                  type: string
                  example: 117.png
                interest:
                  type: number
                  example: 5
      summary: 유저의 프로필 변경
      description: 유저의 프로필 변경 API
      security:
        - Authorization: []
      responses:
        '200':
          description: 유저 프로필 변경 성공
  /user/changePassword:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  example: 1234
      tags:
        - user
      summary: 비밀번호 변경
      description: User의 비밀번호 변경 API
      security:
        - Authorization: []
      responses:
        '200':
          description: 비밀번호 변경 성공
        '400':
          description: 변경할 패스워드 입력
...
  '/pokemon/{id}':
    get:
      summary: 포켓몬 모든 정보 반환
      description: |
        포켓몬의 id를 params로 받아 해당 포켓몬의 정보를 반환
      tags:
        - pokemon
      security:
        - Authorization: []
      parameters:
        - in: path
          name: id
          required: true
          type: number
          schema:
            type: number
            maximum: 151
            examples:
              zero:
                value: 0
              max:
                value: 151
          style: matrix
      responses:
        '200':
          description: 포켓몬 정보 요청 정상 처리
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                  id:
                    type: number
                  name:
                    type: string
                  status:
                    type: string
                  typeOne:
                    type: string
                  typeTwo:
                    type: string
                  height:
                    type: number
                  weight:
                    type: number
                  totalPoints:
                    type: number
                  hp:
                    type: number
                  attack:
                    type: number
                  defense:
                    type: number
                  spAttack:
                    type: number
                  spDefense:
                    type: number
                  speed:
                    type: number
              example:
                _id: '626407bd961c4243634ea1e8,'
                id: 2
                name: 이상해풀
                status: 노말
                typeOne: 풀
                typeTwo: 독
                height: 1
                weight: 13
                totalPoints: 405
                hp: 60
                attack: 62
                defense: 63
                spAttack: 80
                spDefense: 80
                speed: 60
        '400':
          description: 요청이 잘못 됨
  '/pokemon/{id}/name':
    get:
      summary: '포켓몬 id, 이름 반환'
      description: |
        포켓몬의 id를 params로 받아 해당 포켓몬의 id와 이름(name)을 반환
      tags:
        - pokemon
      security:
        - Authorization: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
            maximum: 151
            examples:
              zero:
                value: 0
              max:
                value: 151
          style: matrix
      responses:
        '200':
          description: '포켓몬 id, name 요청 정상 처리'
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                  name:
                    type: string
              example:
                id: 2
                name: 이상해풀
        '400':
          description: 요청이 잘못 됨
  /pokemonList:
    get:
      summary: 모든 포켓몬의 정보를 리스트로 반환
      description: |
        각각 포켓몬의 모든 정보를 담은 object들을 array에 담아 모든 포켓만을 반환
      tags:
        - pokemon
      security:
        - Authorization: []
      responses:
        '200':
          description: 포켓몬들 정보 요청 정상 처리
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    _id:
                      type: string
                    id:
                      type: number
                    name:
                      type: string
                    status:
                      type: string
                    typeOne:
                      type: string
                    typeTwo:
                      type: string
                    height:
                      type: number
                    weight:
                      type: number
                    totalPoints:
                      type: number
                    hp:
                      type: number
                    attack:
                      type: number
                    defense:
                      type: number
                    spAttack:
                      type: number
                    spDefense:
                      type: number
                    speed:
                      type: number
                example:
                  - _id: 626407bd961c4243634ea1e7
                    id: 1
                    name: 이상해씨
                    status: 노말
                    typeOne: 풀
                    typeTwo: 독
                    height: 0.7
                    weight: 6.9
                    totalPoints: 318
                    hp: 45
                    attack: 49
                    defense: 49
                    spAttack: 65
                    spDefense: 65
                    speed: 45
                  - _id: '626407bd961c4243634ea1e8,'
                    id: 2
                    name: 이상해풀
                    status: 노말
                    typeOne: 풀
                    typeTwo: 독
                    height: 1
                    weight: 13
                    totalPoints: 405
                    hp: 60
                    attack: 62
                    defense: 63
                    spAttack: 80
                    spDefense: 80
                    speed: 60
                  - _id: 626407bd961c4243634ea1e9
                    id: 3
                    name: 이상해꽃
                    status: 노말
                    typeOne: 풀
                    typeTwo: 독
                    height: 2
                    weight: 100
                    totalPoints: 525
                    hp: 80
                    attack: 82
                    defense: 83
                    spAttack: 100
                    spDefense: 100
                    speed: 80
                  - _id: '626407bd961c4243634ea1ea,'
                    id: 4
                    name: 파이리
                    status: 노말
                    typeOne: 불꽃
                    typeTwo: 없음
                    height: 0.6
                    weight: 8.5
                    totalPoints: 309
                    hp: 39
                    attack: 52
                    defense: 43
                    spAttack: 60
                    spDefense: 50
                    speed: 65
        '400':
          description: 요청이 잘못
  '/pokemonList/{type}':
    get:
      summary: 특정 속성 포켓몬의 정보를 리스트로 반환
      description: |
        각각 포켓몬의 모든 정보를 담은 object들을 array에 담아 특정 속성의 포켓만을 반환.<br> typeOne, typeTwo 둘 중 하나라도 params type과 일치하면 반환
      tags:
        - pokemon
      security:
        - Authorization: []
      parameters:
        - in: path
          name: type
          required: true
          schema:
            type: string
            enum:
              - 풀
              - 불꽃
              - 물
              - 벌레
              - 노말
              - 독
              - 전기
              - 땅
              - 페어리
              - 격투
              - 에스퍼
              - 바위
              - 고스트
              - 얼음
              - 드래곤
              - 비행
              - 강철
          style: matrix
      responses:
        '200':
          description: 포켓몬들 정보 요청 정상 처리
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    _id:
                      type: string
                    id:
                      type: number
                    name:
                      type: string
                    status:
                      type: string
                    typeOne:
                      type: string
                    typeTwo:
                      type: string
                    height:
                      type: number
                    weight:
                      type: number
                    totalPoints:
                      type: number
                    hp:
                      type: number
                    attack:
                      type: number
                    defense:
                      type: number
                    spAttack:
                      type: number
                    spDefense:
                      type: number
                    speed:
                      type: number
                example:
                  - typeTwo: 없음
                    _id: 626407bd961c4243634ea1ff
                    id: 25
                    name: 피카츄
                    status: 노말
                    typeOne: 전기
                    height: 0.4
                    weight: 6
                    totalPoints: 320
                    hp: 35
                    attack: 55
                    defense: 40
                    spAttack: 50
                    spDefense: 50
                    speed: 90
                  - typeTwo: 없음
                    _id: 626407bd961c4243634ea24b
                    id: 101
                    name: 붐볼
                    status: 노말
                    typeOne: 전기
                    height: 1.2
                    weight: 66.6
                    totalPoints: 490
                    hp: 60
                    attack: 50
                    defense: 70
                    spAttack: 80
                    spDefense: 80
                    speed: 150
                  - _id: 626407bd961c4243634ea238
                    id: 82
                    name: 레어코일
                    status: 노말
                    typeOne: 전기
                    typeTwo: 강철
                    height: 1
                    weight: 60
                    totalPoints: 465
                    hp: 50
                    attack: 60
                    defense: 95
                    spAttack: 120
                    spDefense: 70
                    speed: 70
        '400':
          description: 요청이 잘못
  /drawPokemon:
    get:
      summary: 포켓몬 스티커 뽑기
      description: 포켓몬 스티커를 등급에 따라 확률적으로 뽑아주고 업적 및 랭크를 갱신
      tags:
        - pokemon
      security:
        - Authorization: []
      responses:
        '200':
          description: |
            포켓몬 스티커 뽑기 및 업적•랭크 포인트 갱신 완료.<br> 뽑힌 포켓몬의 정보와 유저 잔여 포인트, 랭킹 포인트를 함께 반환.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                  name:
                    type: string
                  pokemonStatus:
                    type: string
                  pokemonTotalPoint:
                    type: number
                  status:
                    type: boolean
                    description: 성공 여부
                  userPoint:
                    type: number
                  PankingPoint:
                    type: number
              example:
                id: 132
                name: 메타몽
                pokemonStatus: 노말
                pokemonTotalPoint: 288
                status: true
                userPoint: 4500
                RankingPoint: 656.5
        '400':
          description: 요청이 잘못
...